requirements.txt
anyio==3.6.1
bidict==0.22.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
websocket-client==1.4.1
websockets==10.3
